requirements.txt
Django==1.10.4
django-guardian==1.4.9
django-oauth-toolkit==0.12.0
djangorestframework==3.6.3